third_party/blink/web_tests/external/wpt/css/cssom-view/element-scroll-promises-expected.txt

This is a testharness.js-based test.
[FAIL] Element.scrollIntoView(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
Harness: the test ran to completion.
